functionalities
timeline, onion skin and particles prep

+++ src/main/java/net/pulga22/particlestudio/core/routines/ParticlePoint.java
@@ -4,10 +4,10 @@ import java.io.Serializable;
 
 public class ParticlePoint implements Serializable {
 
-    public final float[] position = {0, 0, 0};
+    public final double[] position = {0, 0, 0};
     public final String particleType;
 
-    public ParticlePoint(String particleType, float x, float y, float z){
+    public ParticlePoint(String particleType, double x, double y, double z){
         position[0] = x;
         position[1] = y;
         position[2] = z;
